Make KickoffActivity truly transparent
