+++ auth/src/main/java/com/firebase/ui/auth/ui/AppCompatBase.java
@@ -14,31 +14,17 @@
 
 package com.firebase.ui.auth.ui;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.RestrictTo;
-import android.support.v7.app.AppCompatActivity;
 
 @SuppressWarnings("Registered")
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-public class AppCompatBase extends AppCompatActivity {
-
-    protected ActivityHelper mActivityHelper;
+public class AppCompatBase extends HelperActivityBase {
 
     @Override
     protected void onCreate(Bundle savedInstance) {
         super.onCreate(savedInstance);
-        mActivityHelper = new ActivityHelper(this, getIntent());
         mActivityHelper.configureTheme();
     }
 
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        mActivityHelper.dismissDialog();
-    }
-
-    public void finish(int resultCode, Intent intent) {
-        mActivityHelper.finish(resultCode, intent);
-    }
 }
