+++ auth/src/main/res/values/styles.xml
@@ -15,6 +15,15 @@
         <item name="colorControlHighlight">@color/authui_colorAccent</item>
     </style>
 
+    <style name="FirebaseUI.Transparent" parent="FirebaseUI">
+        <item name="android:windowIsTranslucent">true</item>
+        <item name="android:windowBackground">@android:color/transparent</item>
+        <item name="android:windowContentOverlay">@null</item>
+        <item name="android:windowNoTitle">true</item>
+        <item name="android:windowIsFloating">true</item>
+        <item name="android:backgroundDimEnabled">false</item>
+    </style>
+
     <style name="FirebaseUI.Dialog" parent="Base.Theme.AppCompat.Light.Dialog"/>
 
     <style name="FirebaseUI.WrapperStyle">
